Updated session based conditional auth functions according to the changes requested in the reviews
