+++ components/org.wso2.carbon.identity.conditional.auth.functions.session/src/main/java/org/wso2/carbon/identity/conditional/auth/functions/session/model/Session.java
@@ -31,6 +31,9 @@ public class Session {
     private String ipAddress;
     private String serviceProvider;
 
+    public String getSessionId() {
+        return sessionId;
+    }
     /**
      * Constructor of the model class.
      *
